construct multiple sections
