+++ conf/messages.en
@@ -201,3 +201,4 @@ viewReturnSummary.summary.field.1 = Tax liability for this period
 viewReturnSummary.summary.field.2 = Return submitted
 viewReturnSummary.summary.field.3 = Payment due
 viewReturnSummary.summary.field.4 = Return reference number
+viewReturnSummary.details.heading = Details
